requirements.txt
asgiref==3.5.2
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.0.12
Django==4.0.5
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-rest-framework==0.1.0
django-six==1.0.4
djangorestframework==3.13.1
idna==3.3
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
python-decouple==3.6
pytz==2022.1
requests==2.28.0
screen==1.0.1
SQLAlchemy==1.4.39
sqlparse==0.4.2
tomli==2.0.1
twilio==7.9.3
urllib3==1.26.9
xlwt==1.3.0
